Create ShimmerLoader.dart
Subject: Added ShimmerLoader widget

Hi CodeBooter,

I'm submitting a pull request to add a new ShimmerLoader widget to CodeBooter. The widget is a generalized shimmer loader that can be used in any Flutter app. It allows you to easily apply a shimmering effect to any widget, such as an image, text, or a container.

The code is very easy to use. Simply wrap the widget that you want to apply the shimmer effect to with the ShimmerLoader widget. You can also customize the shimmer effect by passing in a different LinearGradient object and adjusting the width and height of the shimmer effect.

This code can be used in a variety of ways, such as:

To indicate that data is loading
To create a placeholder for an image or other asset that has not yet loaded
To create a visually appealing effect
I've also added a few unit tests to ensure that the widget works as expected.

Please let me know if you have any questions or feedback.

Thanks,
Atharva-Werulkar